Remove calls to cudaFuncGetAttributes and cudaOccupancyMaxPotentialBlockSize from constructor
The calls to cudaFuncGetAttributes and cudaOccupancyMaxPotentialBlockSize /
cudaOccupancyMaxPotentialBlockSizeVariableSMem needed to be removed from
from the constructor.
This is necessary because cuda::function s are typically global objects whos
constructor will be called *before* the correct device is selected.
Becuase CUDA just selects the first available device when no context was
set the calls could have returned results for the wrong device.

The call to cudaFuncGetAttributes will now be done once when a function
requesting an attribute is called.

Calls to cudaOccupancyMaxPotentialBlockSize and
cudaOccupancyMaxPotentialBlockSizeVariableSMem can not be made from "normal"
(not CUDA) C++ code (https://docs.nvidia.com/cuda/cuda-runtime-api/group__CUDART__HIGHLEVEL.html#group__CUDART__HIGHLEVEL).

Because of this the occupancy function is put into a lambda
expression which will be assigned to the new member occupancy_ of type
std::function.
The methods min_grid_size and max_block_size then call occupancy_ once when
the value is requested.