Add iterator tag for managed memory
This commit adds a random access iterator tag for managed memory.
This was necessary because managed memory can be seen both as host and
device memory and therefore the iterator tag needed to be convertible to
both the host and device iterator tag.
The cuda::copy functions needed to be changed too because the managed
memory iterator tag would lead to ambiguity.

Also add a new iterator test.