Don't use std::vector as base class for managed memory vector
Using std::vector as base class could lead to problems because the base class
accessed memory which lead to the data being migrated back to the host. This
could lead to very bad performance because the unified memory system needed
to copy the data back and forth between the host and the device even if the
data was only accessed from the device.