test: Add REQUIRED_FILES property to tests using cuda-memcheck
If the test executable was not to be created for whatever reason
the tests would still be run (and fail) because the actual executable
being executed by ctest is cuda-memcheck. By adding the REQUIRED_FILES
property the test depends on the correct executable and will only be
run when it is also created.